chore: refactor time range module
